+++ code/headers/module.hpp
@@ -2,6 +2,7 @@
 
 #include <base_module.hpp>
 #include <hwlib.hpp>
+#include <uart_nmea_c.hpp>
 
 namespace r2d2::location_detector {
 
@@ -11,10 +12,13 @@ namespace r2d2::location_detector {
     /// TODO: what is the input (frame) of this module?
     /// what is the output (frame) of this module?
     class module_c : public base_module_c {
+    private:
+       r2d2::location_detector::uart_nmea_c &nmea;
+
     public:
         /// \brief
         /// location_detector module_c constructor.
-        module_c(base_comm_c &comm) : base_module_c(comm) {
+        module_c(r2d2::base_comm_c &comm, r2d2::location_detector::uart_nmea_c &nmea) : base_module_c(comm), nmea(nmea) {
             // TODO: what frames are we listening for?
         }
 
@@ -26,22 +30,7 @@ namespace r2d2::location_detector {
         void process();
 
     protected:
-        /**
-         * \brief Pure virtual interface function to get the location.
-         */
 
-        virtual frame_coordinate_s get_location() = 0;
-        /**
-         * \brief This method is used to manually make coordinate frames.
-         *
-         * \param longitude
-         * \param latitude
-         * \param north
-         * \param south
-         * \param altitude
-         */
-        frame_coordinate_s compress(float longitude, float latitude, bool north,
-                                    bool east, int16_t altitude);
     };
 
 } // namespace r2d2::location_detector
